requirements.txt
Django~=4.0.4
PyMySQL==1.0.3
dialogflow==0.5.1
djangorestframework~=3.13.1
pip~=22.0.4
wheel~=0.37.1
setuptools~=62.1.0
pytz~=2022.1
sqlparse~=0.4.2